requirements.txt
agora-token-builder==1.0.0
ansicon==1.89.0
arrow==1.2.1
asgiref==3.5.0
backports.entry-points-selectable==1.1.0
bcrypt==3.2.0
blessed==1.19.0
cffi==1.15.0
cryptography==36.0.1
distlib==0.3.3
Django==4.0.2
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-filter==21.1
django-picklefield==3.0.1
django-q==1.3.9
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
filelock==3.3.1
gunicorn==20.1.0
jinxed==1.1.0
librouteros==3.2.0
paramiko==2.8.1
Pillow==9.0.1
platformdirs==2.4.0
pycparser==2.21
PyJWT==2.3.0
PyNaCl==1.4.0
python-dateutil==2.8.2
pytz==2022.1
redis==3.5.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
virtualenv==20.9.0
wcwidth==0.2.5
whitenoise==5.3.0
